[ORC] Lift SymbolLookupSet into its own header. NFCI.
Lifts the SymbolLookupSet class and SymbolLookupFlags enum out of Core.h and
into a new SymbolLookupSet.h header.
